refactor(search): optimize client filtering with useMemo and simplify state
